Make new gqlMock input and move all into new file.
