+++ services/app-api/src/domain-models/contractAndRates/index.ts
@@ -47,3 +47,5 @@ export type {
     ContractPackageSubmissionType,
     ContractPackageSubmissionWithCauseType,
 } from './packageSubmissions'
+
+export { validateContractDraftRevisionInput } from './dataValidatorHelpers'
